+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - ...
+### Added
+- Script for downloading test images.
+
 ## [0.3.1] - 2022-09-09
 ### Fixed
 - Support for openslide-python 1.2.