+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Script for downloading test images.
 
+### Fixed
+- Handling of OpenSlide images without slide background color.
+
 ## [0.3.1] - 2022-09-09
 ### Fixed
 - Support for openslide-python 1.2.