+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] -
 
+## [0.5.0] - 2022-01-16
+
 ### Changed
 
 - include_levels now takes a list of level indices instead of pyramid levels.
@@ -94,7 +96,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of wsidicomizer
 
-[Unreleased]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.4.0..HEAD
+[Unreleased]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.5.0..HEAD
+[0.5.0]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.4.0..0.5.0
 [0.4.0]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.3.1..0.4.0
 [0.3.1]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.3.0..0.3.1
 [0.3.0]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.2.0..0.3.0